Testable version + TorrentClient bugfix.
This version can now be run: waf now produces a userInterface
executable which, when run, will request the path to a torrent
file, the path to the directory to download/upload from,
and it will attempt to download/upload the file.

Currently crashes - crashes are not always the same -
can crash for multiple different reasons.  Crashes both for
uploads and downloads.

NOTE: mangled the wscript, now unit tests no longer automagically
run upon building.